feat(deploy): refresh runtime and add production stack (#4)
* Fix compiler and frontend warnings

* Refresh frontend dependencies

* Refresh Rust lockfile

* Update tracing subscriber

* Update validator dependency

* Update jsonwebtoken dependency

* Update Rust CLI dependencies

* Fix Clap hostname flag conflict

* Update small Rust dependencies

* Update reqwest dependency

* Remove unused server dependencies

* Update SQLx dependency

* Update Wasmtime dependency

* Refresh Rust patch dependencies

* Update jsonwebtoken dependency

* Trim Ramiel logging dependency features

* feat(deploy): add production stack

* fix(security): harden production runtime

* fix(runner): harden compilation runtime

* fix(runner): address review follow-ups

* fix(deploy): harden Caddy runtime

---------

+++ crates/server/src/user/star_count.rs
@@ -6,7 +6,7 @@ use crate::error::{ServerError, UserError};
 
 #[derive(Serialize)]
 pub struct StarCount {
-    count: i32,
+    count: i64,
 }
 
 pub async fn star_count(
